Add get_or_insert method
We don't follow the entry() API exposed by BTreeMap because that's a bit
tricky to implement; the rebalancing logic in TreeMap wants to run after
we've inserted the element (if we're going to), so with Entry we don't
know exactly when to do the rebalancing if the client doesn't insert
or remove via the Entry. This method here is simpler because we know
we'll always insert an element if there isn't one, and knowing this is
very common.